+++ packages/core/test/render3/perf/ng_template/index.ts
@@ -6,15 +6,16 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import {injectTemplateRef} from '@angular/core/src/linker/template_ref';
+import {injectViewContainerRef} from '@angular/core/src/linker/view_container_ref';
 import {TemplateRef, ViewContainerRef} from '../../../../src/linker';
 import {ɵɵdefineDirective, ɵɵdirectiveInject, ɵɵtemplate} from '../../../../src/render3/index';
 import {createLView, createTNode, createTView} from '../../../../src/render3/instructions/shared';
 import {RenderFlags} from '../../../../src/render3/interfaces/definition';
 import {TNodeType} from '../../../../src/render3/interfaces/node';
 import {LViewFlags, TViewType} from '../../../../src/render3/interfaces/view';
-import {injectViewContainerRef} from '../../../../src/render3/view_engine_compatibility';
 import {createBenchmark} from '../micro_bench';
 import {createAndRenderLView} from '../setup';
+
 
 
 class TemplateRefToken {
@@ -32,7 +33,7 @@ class ViewContainerRefToken {
    * @nocollapse
    */
   static __NG_ELEMENT_ID__(): ViewContainerRef {
-    return injectViewContainerRef(ViewContainerRef);
+    return injectViewContainerRef();
   }
 }
